feat: Extracting find packakge function
